notify: skip merged pr statuses for now, and skip already automated users
These changes are a temporary test, since there is a lot of noise from
all the pr's we want to try and skip the merged pr statuses since they
might not be used that much. PT might be a better overview, especially
if work is split accross multiple pr's.
Also removing duplicate messages about pr's in these volatile messages,
since these are automated by the users themself which allows threaded discussion to remain.